+++ pystrich/datamatrix/renderer.py
@@ -1,17 +1,25 @@
 """Datamatrix renderer"""
 
-__revision__ = "$Rev$"
+from __future__ import annotations
 
+import os
 from io import BytesIO
+from typing import TYPE_CHECKING
 
 from PIL import Image
 
 from pystrich.exceptions import PyStrichInvalidOption
 
+if TYPE_CHECKING:
+    from PIL.Image import Image as PILImage
+
 DATAMATRIX_DEFAULT_QUIET_ZONE = 2
 
+_PIXEL: dict[int | None, bytes] = {0: b"\xff", 1: b"\x00"}
+_SYMBOL: dict[int | None, str] = {0: '  ', 1: 'XX'}
 
-def repr_matrix(matrix):
+
+def repr_matrix(matrix: list[list[int | None]]) -> str:
     return "\n".join(repr(x) for x in matrix)
 
 
@@ -20,7 +28,20 @@ class DataMatrixRenderer:
     it will add edge handles and render to either to an image
     (including quiet zone) or ascii printout"""
 
-    def __init__(self, matrix, regions, *, quiet_zone=DATAMATRIX_DEFAULT_QUIET_ZONE):
+    width: int
+    height: int
+    regions: int
+    region_size: int
+    quiet_zone: int
+    matrix: list[list[int | None]]
+
+    def __init__(
+        self,
+        matrix: list[list[int | None]],
+        regions: int,
+        *,
+        quiet_zone: int = DATAMATRIX_DEFAULT_QUIET_ZONE,
+    ) -> None:
         self.width = len(matrix)
         self.height = len(matrix[0])
         self.regions = regions
@@ -37,13 +58,13 @@ class DataMatrixRenderer:
         # add the edge handles
         self.add_handles()
 
-    def put_cell(self, position, colour=1):
+    def put_cell(self, position: tuple[int, int], colour: int = 1) -> None:
         """Set the contents of the given cell"""
 
         posx, posy = position
         self.matrix[posy][posx] = colour
 
-    def add_handles(self):
+    def add_handles(self) -> None:
         """Set up the edge handles"""
 
         for x_index in range(self.regions):
@@ -69,7 +90,7 @@ class DataMatrixRenderer:
                 for i in range(y_max, y_origin, -2):
                     self.put_cell((x_max, i))
 
-    def add_border(self, colour=1):
+    def add_border(self, colour: int = 1) -> None:
         """Wrap the matrix in a border of given width
             and colour"""
 
@@ -77,7 +98,7 @@ class DataMatrixRenderer:
         self.width += a_gap*2 + self.quiet_zone*2 + (self.regions-1)*a_gap*2
         self.height += a_gap*2 + self.quiet_zone*2 + (self.regions-1)*a_gap*2
 
-        new_matrix = []
+        new_matrix: list[list[int | None]] = []
         for i in range(a_gap+self.quiet_zone):
             new_matrix += [[colour]*self.width]
 
@@ -87,7 +108,7 @@ class DataMatrixRenderer:
                 for j in range(a_gap*2):
                     new_matrix += [[colour]*self.width]
             # Left gap
-            new_row = [colour]*(a_gap+self.quiet_zone)
+            new_row: list[int | None] = [colour]*(a_gap+self.quiet_zone)
             # Split according to regions
             for i in range(self.regions):
                 part = row[i*self.region_size:(i+1)*self.region_size]
@@ -103,7 +124,7 @@ class DataMatrixRenderer:
             new_matrix += [[colour]*self.width]
         self.matrix = new_matrix
 
-    def get_pilimage(self, cellsize):
+    def get_pilimage(self, cellsize: int) -> PILImage:
         """Return the matrix as an PIL object"""
 
         # get the matrix into the right buffer format
@@ -115,51 +136,36 @@ class DataMatrixRenderer:
                                buff, 'raw', 'L', 0, -1)
         return img
 
-    def write_file(self, cellsize, filename):
+    def write_file(self, cellsize: int, filename: str | os.PathLike[str]) -> None:
         """Write the matrix out to an image file"""
         img = self.get_pilimage(cellsize)
         img.save(filename)
 
-    def get_imagedata(self, cellsize):
+    def get_imagedata(self, cellsize: int) -> bytes:
         """Write the matrix out as PNG to an bytestream"""
         imagedata = BytesIO()
         img = self.get_pilimage(cellsize)
         img.save(imagedata, "PNG")
         return imagedata.getvalue()
 
-    def get_buffer(self, cellsize):
+    def get_buffer(self, cellsize: int) -> bytes:
         """Convert the matrix into the buffer format used by PIL"""
-
-        def pixel(value):
-            """return pixel representation of a matrix value
-            0 => white, 1 => black"""
-            if value == 0:
-                return b"\xff"
-            elif value == 1:
-                return b"\x00"
 
         # PIL writes image buffers from the bottom up,
         # so feed in the rows in reverse
         buf = b""
         for row in self.matrix[::-1]:
-            bufrow = b''.join([pixel(cell) * cellsize for cell in row])
+            bufrow = b''.join([_PIXEL[cell] * cellsize for cell in row])
             for _ in range(0, cellsize):
                 buf += bufrow
         return buf
 
-    def get_ascii(self):
+    def get_ascii(self) -> str:
         """Write an ascii version of the matrix out to screen"""
 
-        def symbol(value):
-            """return ascii representation of matrix value"""
-            if value == 0:
-                return '  '
-            elif value == 1:
-                return 'XX'
+        return '\n'.join(''.join(_SYMBOL[cell] for cell in row) for row in self.matrix) + '\n'
 
-        return '\n'.join([''.join([symbol(cell) for cell in row]) for row in self.matrix]) + '\n'
-
-    def get_dxf(self, cellsize, inverse, units):
+    def get_dxf(self, cellsize: float, inverse: bool, units: str) -> str:
         """Write an DXF version of the matrix to a string"""
         dxf = []
         dxf.append("0\nSECTION\n2\nHEADER\n")